all right

+++ include/Types.h
@@ -5,8 +5,8 @@
 #ifndef PAA_FIRST_ASSIGNMENT_TYPES_H
 #define PAA_FIRST_ASSIGNMENT_TYPES_H
 
-typedef long long int v_T;       //type of vertices
-typedef long long int e_T;       //type of the label to edge
-typedef long long w_T;           //weight
+typedef long int v_T;       //type of vertices
+typedef long int e_T;       //type of the label to edge
+typedef long long int w_T;           //weight
 
 #endif //PAA_FIRST_ASSIGNMENT_TYPES_H
